@@ -1,5 +1,5 @@
-using System.ComponentModel.DataAnnotations;
 using System.Collections.Generic;
+using System.ComponentModel.DataAnnotations;
 
 namespace WebHackingPortal.Models
 {
@@ -8,26 +8,17 @@ namespace WebHackingPortal.Models
         [Key]
         public int Id { get; set; }
 
-        [Required]
         public string Title { get; set; } = string.Empty;
-
-        [Required]
         public string Description { get; set; } = string.Empty;
-
-        [Required]
-        public string Category { get; set; } = string.Empty;
-
-        public int Points { get; set; }
-
-        [Required]
         public string Flag { get; set; } = string.Empty;
 
+        public int Points { get; set; }
         public int Order { get; set; }
 
-        public int? LabId { get; set; }
+        // Foreign key to Lab
+        public int LabId { get; set; }
         public Lab? Lab { get; set; }
 
-        // Navigation properties
         public ICollection<Submission> Submissions { get; set; } = new List<Submission>();
     }
 }
