@@ -1,5 +1,6 @@
 using System;
 using System.ComponentModel.DataAnnotations;
+using System.ComponentModel.DataAnnotations.Schema;
 
 namespace WebHackingPortal.Models
 {
@@ -8,17 +9,18 @@ namespace WebHackingPortal.Models
         [Key]
         public int Id { get; set; }
 
-        public int UserId { get; set; }
-        public AppUser? User { get; set; }
-
-        public int ChallengeId { get; set; }
-        public Challenge? Challenge { get; set; }
-
         [Required]
+        [ForeignKey("User")]
+        public int UserId { get; set; }
+        public int LabId { get; set; }
+        public int ChallengeId { get; set; }
+
         public string SubmittedFlag { get; set; } = string.Empty;
-
         public bool IsCorrect { get; set; }
-
         public DateTime Timestamp { get; set; } = DateTime.UtcNow;
+
+        // Navigation properties (optional)
+        [InverseProperty("Submissions")]
+        public AppUser? User { get; set; }
     }
 }
